separate data extension class from detector constructor ready to be upstreamed to DD4hep
